+++ app/controllers/conversatio/posts_controller.rb
@@ -4,7 +4,7 @@ class Conversatio::PostsController < ApplicationController
 
   def show
     @blog = Blog.find params[:blog_id]
-    @post = @blog.published_posts.find params[:id]
+    @post = @blog.published_posts.until_now.select{|i| i==Post.find(params[:id])}.first
   end
 
   def by_tag
@@ -38,10 +38,10 @@ class Conversatio::PostsController < ApplicationController
     to   = DateTime.new(to.year, to.month, -1, 23, 59, 59) if @day.nil?
 
     @blog = Blog.find params[:blog_id]
-    @posts = @blog.published_posts.paginate :per_page => 10,
-                                            :page => @page,
-                                            :conditions => ['published_at between ? and ?', from, to],
-                                            :order => "published_at desc"
+    @posts = @blog.published_posts.until_now.paginate :per_page => 10,
+                                                                :page => @page,
+                                                                :conditions => ['published_at between ? and ?', from, to],
+                                                                :order => "published_at desc"
   end
 
 end